Split the modulemd parser into its own file
